Added status column to admin panel
