+++ src/util/types.ts
@@ -9,3 +9,16 @@ export interface RegistrationForm {
   username: string;
   password: string;
 }
+
+type StatusKeys = "accepted" | "pending" | "rejected";
+
+/**
+ * Specifies the statuses an application be set to by the recruiter.
+ */
+export type Statuses = {
+  [status in StatusKeys]: {
+    i18nPath: string;
+    icon: string;
+    color: string;
+  };
+};
